Correctly handle 'Copy to Chrome' when Chrome is running in multitasking mode.
This CL fixes 'Copy to Chrome' in multitasking mode when
kExternalFilesLoadedInWebState flag is enabled.

URL is set to completeURL (which is file:// URL) and virtualURL is
set to externalURL (which is chrome:// URL). This is necessary because
WebState knows how to render file URLs but can not render chrome:// URL.

The code was copied from UserAcrtivityHandler, so this CL can be
cherry-picked to M73. Code duplication cleanup will be done in a follow
up CL.

Bug: 934789
Change-Id: I490cea33171b25ff73ef713c8610a64da9d8145f
Reviewed-on: https://chromium-review.googlesource.com/c/1484631
Commit-Queue: Mark Cogan <marq@chromium.org>
Auto-Submit: Eugene But <eugenebut@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#635222}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 998c1103ecf3e86b16451a6272562d16edad429c