[iOS] Stop Voiceover reading items on the WebView when omnibox is open
Currently, when the omnibox popup is open, Voiceover will highlight
and read the items from the WebView, even though those views are
covered by the popup. This is because the popup is just added to the
same parent VC (not presented). We need something to manually say that
the webview is covered.

Bug: 786940, 912419
Change-Id: I70ce9abcfff8c3f0e48aa990d383eabe78b6ef47
Reviewed-on: https://chromium-review.googlesource.com/c/1485951
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#636754}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f56c7f032fcadc55fd068461f4aef75955974c9