[ios] Displays InfobarBadge when an Infobar is presented.
LocationBar
- Sets the LocationBarMediator as a InfobarBadgeTabHelper delegate.
- Adds a LocationBarConsumer method to communicate the badge displaying
from the Mediator to the ViewController.

Infobar
- Moves self.delegate->RemoveInfoBar(); from stop to deatachView to
prevent the Infobar to be destroyed when switching WebStates. This
way the badge will persist when returning to a WebState that was
previously displaying a badge.
- Adds InfobarBadgeDisplaying() to InfobarBadgeTabHelper so the Delegate
knows if the newWebState should display a badge or not.

Bug: 911864
Change-Id: I3da2898e1597900da79b83b17e75c210c420516a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1493388
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#637386}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5be178683fd0112813f23a94449df64afb0eb628