[iOS][MF] Use city and email as title fallbacks
This changes the logic of the title for the Manual Fallback addresses
cells. Previosly the line 1 was always in black and the line 2 was in
gray.

Now the logic is to fallback in the following order:
1. Line 1 and Line 2 (as previously)
2. Line 2
3. City
4. Email

Bug: 924097
Change-Id: Id8b4c7d0b69729c5b2f0008868341d690697256c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1503312
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#637654}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d1d7d763f8d99dee080c252688b1a31b06855efb