[iOS][web] Move helper methods around
This CL moves the helper methods around in CRWWebController.
It moves methods that were in a pragma section to a unnamed section if
it didn't make sense for them to be the previous section.
It is also starting to move the comments for private methods to the
implementation.

Bug: 938255
Change-Id: I8408fe3762f7de59c10eeccd009c70a04425e97d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1503457
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#638135}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 275f98f155464fbc8632ee3f2a431f8b36cd5ea9