Eliminate FakeAccountFetcherService
This CL eliminates FakeAccountFetcherService, which provides no
functionality over AccountFetcherService following recent changes.

The TestImageDecoder class that was contained in
fake_account_fetcher_service.* is retained, as it has usage in the
codebase. It looks like that usage could potentially be replaced by
usage of FakeImageDecoder, but that project is orthogonal to the
project of eliminating FakeAccountFetcherService.

TBR=droger@chromium.org

Bug: 935998
Change-Id: I4eed6e8f4b3a9474076eeed27e7641cd88b25458
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1503331
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#638551}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da96e6ab65c029e04bbf95f7fa120d4cbb40bbfe