Call StopLoading before loading data
|CRWWebController loadData:| will replace the content
of the page loaded by the lastCommittedItem.
If there is a pending item, this pending loading must
be cancelled before proceeding in loading data.

This can be the case when loading offline page in place
of an uncommitted reloading navigation.

Bug: 942025
Change-Id: Id735961110218eb9ede2531c19c33b29879a688f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1524264
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#641190}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 88cd402005df5075e1ee389d001e599cfa1f4a31