[iOS] Move SizeClass to WelcomeToChrome
This CL moves the content of the UIView+SizeClassSupport to
WelcomeToChromeView as it was the only class using it and there isn't
much value to use it in other places.
WelcomeToChromeView should probably be refactored to stop using it.

Bug: 944008
Change-Id: I3a8420538f637084996161f27c1f2833afe3584b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1530873
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#642533}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d3a08d3ad69370b53d937bcd4ca5f229a6a710c