[omnibox] Set accessibility text on revert
Before, the omnibox accessibility text was set when
OnTemporaryTextMaybeChanged was called, which was not called when the
omnibox was reverted. This was fine, as screen readers were only
triggered when traversing the omnibox with the arrow/page up/down keys,
and not when reverting the omnibox. crrev.com/c/1419257 changed
traversing the selection to the first suggestion to be consistent with
reverting the omnibox. Consequently, the accessibility text was no
longer set when arrowing/paging up to the first suggestion, and the
incorrect text (corresponding to the previously selected match) was
read.

This CL sets the accessibility text on revert as well.

Bug: 943105
Change-Id: I22df51b3a17a34d5a7c35f494a365391d1c67236
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1534138
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Kevin Bailey <krb@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#644345}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b29136eff0dee370e1c04255edf6233b257783a