[iOS] Implementing the advance sign-in settings
Adding AdvancedSigninSettingsCoordinator with its view controller.
The goal is to present the sync settings to the user right after signing
in and before to start the sync.
This dialog is presented when the user taps on the setting link in the
sign-in consent.

Screenshots:
https://drive.google.com/open?id=1LA9A2iQR4n0QFhJO5EhBCf8BiAu4qk-5
https://drive.google.com/open?id=1rCawEHHLuzmNRU1EEm5IPBsDPqlv7pLV

Bug: 883075
Change-Id: I0ae187b65b032ff65e569b3568c043c3e7c164bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1539036
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#646285}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 89c33f480b2d90eac564620d08e59a2bcf549743