🦁Update formatting and spacing for omnibox popup cells
This CL contains many layout changes due to UX feedback. More
specifically, it:
- Changes the size of the leading image in the omnibox text field. This
also allows constraining the size of the popup image directly to the
size of the leading image. This also introduces manual constraints for
the image size instead of using the contenti size for whatever image is
in the image view.
- Decreases the offset between the image view and the leading edge in
the omnibox text field.
- Increases the offset between the image and the text in the omnibox
text field.
- Increases the popup row minimum height.
- Increases the trailing margin on the trailing button in the popup row
(for append, switch to open tab).
- Causes the answer image to appear by preventing it from being
compressed horizontally.

It does not make either of the changes to the autocomplete match formatter
(use a bullet for spacing and change the text styling to semibold). Those
changes will come later along with the other changes in autocomplete match
formatter behind a flag.

Bug: 945323
Change-Id: I71ff8584356caa8afaf094423bdad9957e6e7bb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1561113
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#650249}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f969c385c1f69de72277ab89bf1f0781f56ad9dc